rendu fonctionne, mouvement saccade, pb de collisions

+++ Move.h
@@ -1,4 +1,4 @@
 #pragma once
 
-void applyMovement(std::vector<MovingObject>& objects, int deltaTime);
-void applyMovement(MovingObject& objects, int deltaTime);
+void applyMovement(std::vector<MovingObject>& objects, float deltaTime);
+void applyMovement(MovingObject& objects, float deltaTime);
